requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
tzdata==2024.1
Pillow>=10.0.0
requests==2.31.0
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
djangorestframework==3.14.0
markdown==3.5
django-filter==23.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.5.0
redis==4.6.0
django-debug-toolbar==4.2.0
cryptography==41.0.0

